Move additemstoplayer command to admin and update inventory grant
