+++ commands/adminCommands/additemstoplayer.js
@@ -1,0 +1,22 @@
+// Admin command
+
+const { SlashCommandBuilder } = require('discord.js');
+const { grantItemToPlayer } = require('../../db/inventory');
+
+module.exports = {
+    data: new SlashCommandBuilder()
+        .setName('additemstoplayer')
+        .setDescription('Adds items to a player')
+        .setDefaultMemberPermissions(0)
+        .addUserOption(option => option.setName('player').setDescription('The player to add items to').setRequired(true))
+        .addStringOption(option => option.setName('item').setDescription('The item to add').setRequired(true))
+        .addIntegerOption(option => option.setName('quantity').setDescription('The amount of items to add').setRequired(false)),
+    async execute(interaction) {
+        const player = interaction.options.getUser('player');
+        const item = interaction.options.getString('item');
+        const qty = interaction.options.getInteger('quantity') ?? 1;
+
+        await grantItemToPlayer(player.id, item, qty);
+        return interaction.reply({ content: `Gave ${qty} ${item} to <@${player.id}>`, ephemeral: true });
+    },
+};
